top level await workaround
