+++ .github/workflows/check-milestone.yml
@@ -1,5 +1,5 @@
 name: Automatically set milestone
-run-name: Set milestone for ${{ github.event.head_commit.message }}
+run-name: Set milestone for ${{ github.event.head_commit.message || inputs.version }}
 
 on:
   push:
@@ -16,7 +16,9 @@ jobs:
   set-milestone:
     name: Automatically set milestone
     runs-on: ubuntu-22.04
-    timeout-minutes: 5
+    timeout-minutes: 15
+    env:
+      isNewReleaseBranch: ${{ github.event_name == 'workflow_dispatch' }}
     permissions:
       pull-requests: write
       issues: write
@@ -25,7 +27,7 @@ jobs:
       - uses: actions/checkout@v4
         with:
           sparse-checkout: release
-          fetch-depth: ${{ github.event_name == 'workflow_dispatch' && 0 || 3 }} # we need the full history to get the commit messages
+          fetch-depth: ${{ !env.isNewReleaseBranch && 3 || 0 }} # we need the full history to get all the commit messages
           ref: master # we want the logic from master, even though we're triggering on a release branch
       - name: Prepare build scripts
         run: cd ${{ github.workspace }}/release && yarn && yarn build
@@ -48,8 +50,13 @@ jobs:
               branchName,
               commitMessages: [commitMessage],
             });
+      - name: Get commit messages for release branch
+        if: ${{ env.isNewReleaseBranch }}
+        run: |
+          npm i -g tsx
+          tsx release/src/release-branch-commits.ts ${{ inputs.version }} > commit-messages.json
       - uses: actions/github-script@v7
-        if: ${{ github.event_name == 'workflow_dispatch' }}
+        if: ${{ env.isNewReleaseBranch }}
         name: set milestone for new release branch
         env:
           # this token has a higher rate limit than the default token, and
@@ -57,17 +64,14 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.METABASE_AUTOMATION_USER_TOKEN }}
         with:
           script: | # js
-            const { getReleaseBranchCommits } = require('${{ github.workspace }}/release/dist/index.cjs');
+            const { setMilestoneForCommits } = require('${{ github.workspace }}/release/dist/index.cjs');
+            const commitMessages = require('${{ github.workspace }}/commit-messages.json');
+
             const owner = context.repo.owner;
             const repo = context.repo.repo;
 
             const versionNumber = context.payload.inputs.version;
             const branchName = `release-x.${versionNumber}.x`;
-
-
-            const commitMessages = await getReleaseBranchCommits({
-              versionNumber: Number(versionNumber),
-            });
 
             console.log(`Found ${commitMessages.length} new commits for ${branchName}`);
